+++ site-cookbooks/test-golang/attributes/default.rb
@@ -1,0 +1,3 @@
+default['go']['version'] = '1.7'
+default['go']['gopath'] = '/home/ec2-user/gocode'
+default['go']['gobin'] = '/usr/local/go/bin'