+++ site-cookbooks/test-golang/recipes/default.rb
@@ -1,0 +1,9 @@
+#
+# Cookbook Name:: test-golang
+# Recipe:: default
+#
+# Copyright 2018, YOUR_COMPANY_NAME
+#
+# All rights reserved - Do Not Redistribute
+#
+include_recipe 'golang'