+++ site-cookbooks/test-nginx/attributes/default.rb
@@ -1,0 +1,2 @@
+default['nginx']['conf_template'] = 'isucon_nginx.conf'
+default['nginx']['conf_cookbook'] = 'test-nginx'